Revert "Desugar lib Time API conversions"
This reverts commit 6766facff38ba5eff5a6cda3b3abc7015506a913.

Reason for revert: bots failing

+++ src/test/desugaredLibraryConversions/stubs/Duration.java
@@ -9,7 +9,7 @@ public class Duration {
     return null;
   }
 
-  public long getSeconds() {
+  public int getSeconds() {
     return 0;
   }
